Use system spec for button clicks
